refactor : `orderService` 코드 개선
