+++ src/main/java/kitchenpos/dto/order/OrderRequest.java
@@ -1,37 +1,43 @@
 package kitchenpos.dto.order;
 
+import java.security.InvalidParameterException;
 import java.util.List;
-import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.order.Order;
-import kitchenpos.domain.order.OrderTable;
+import java.util.stream.Collectors;
 
 public class OrderRequest {
 
     private Long orderTableId;
 
-    private OrderLineItemRequests orderLineItems;
+    private List<OrderLineItemRequest> orderLineItems;
 
     public OrderRequest() {
     }
 
     public OrderRequest(Long orderTableId, List<OrderLineItemRequest> orderLineItems) {
         this.orderTableId = orderTableId;
-        this.orderLineItems = new OrderLineItemRequests(orderLineItems);
+        this.orderLineItems = orderLineItems;
     }
 
     public Long getOrderTableId() {
         return orderTableId;
     }
 
-    public OrderLineItemRequests getOrderLineItems() {
+    public List<OrderLineItemRequest> getOrderLineItems() {
         return orderLineItems;
     }
 
     public List<Long> getMenuIds() {
-        return orderLineItems.getMenuIds();
+        return orderLineItems.stream()
+            .map(OrderLineItemRequest::getMenuId)
+            .collect(Collectors.toList());
     }
 
-    public Order toOrder(OrderTable orderTable, List<Menu> menus) {
-        return new Order(orderTable, orderLineItems.toOrderLineItems(menus));
+    public Long getOrderLineItemQuantity(Long menuId) {
+        OrderLineItemRequest orderLineItemRequest = orderLineItems.stream()
+            .filter(orderLineItemRequestTarget -> orderLineItemRequestTarget.isSameMenuId(menuId))
+            .findFirst()
+            .orElseThrow(() -> new InvalidParameterException("메뉴를 찾을 수 없습니다."));
+
+        return orderLineItemRequest.getQuantity();
     }
 }
